Refactor(templates/TS): Update all files after run eslint and add husky
