+++ templates/typescript/.eslintrc.js
@@ -4,62 +4,67 @@ const path = require('path');
 const prettierOptions = JSON.parse(fs.readFileSync(path.resolve(__dirname, '.prettierrc'), 'utf8'));
 
 module.exports = {
-	env: {
-		jest: true,
-		browser: true,
-		node: true,
-		es6: true,
-	},
+  parser: '@typescript-eslint/parser',
 
-	extends: [
-		'eslint:recommended',
-		'plugin:react/recommended',
-		'plugin:@typescript-eslint/recommended',
-		'prettier/@typescript-eslint',
-		'plugin:prettier/recommended',
-	],
+  env: {
+    jest: true,
+    browser: true,
+    node: true,
+    es6: true,
+  },
 
-	globals: {
-		Atomics: 'readonly',
-		SharedArrayBuffer: 'readonly',
-	},
+  extends: [
+    'eslint:recommended',
+    'plugin:react/recommended',
+    'plugin:@typescript-eslint/recommended',
+    'prettier/@typescript-eslint',
+    'plugin:prettier/recommended',
+  ],
 
-	parserOptions: {
-		ecmaFeatures: {
-			jsx: true,
-		},
-		ecmaVersion: 2020,
-		sourceType: 'module',
-	},
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+  },
 
-	plugins: ['react', '@typescript-eslint', 'prettier', 'react-hooks', 'jsx-a11y'],
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+    ecmaVersion: 2020,
+    sourceType: 'module',
+  },
 
-	rules: {
-		'linebreak-style': ['error', 'unix'],
-		'no-console': ['error', { allow: ['warn', 'error'] }],
-		curly: ['error', 'multi'],
-		indent: ['error', 2],
-		quotes: ['error', 'single'],
-		semi: ['error', 'always'],
-		'dot-location': ['error', 'property'],
-		eqeqeq: ['error', 'always'],
-		'react/jsx-filename-extension': 0,
-		'react/jsx-indent': [2, 2, { checkAttributes: true, indentLogicalExpressions: true }],
-		'prettier/prettier': ['error', prettierOptions],
-		'import/no-extraneous-dependencies': 0,
-		// TS
-		'@typescript-eslint/explicit-member-accessibility': 0,
-		'@typescript-eslint/explicit-function-return-type': 0,
-		'@typescript-eslint/triple-slash-reference': 0,
-		'react/no-unescaped-entities': 0,
-	},
+  plugins: ['react', '@typescript-eslint', 'prettier', 'react-hooks', 'import', 'jsx-a11y'],
 
-	settings: {
-		react: {
-			pragma: 'React',
-			version: 'detect',
-		},
-	},
+  rules: {
+    'linebreak-style': ['error', 'unix'],
+    'no-console': ['error', { allow: ['warn', 'error'] }],
+    curly: ['error', 'multi'],
+    indent: ['error', 2],
+    quotes: ['error', 'single'],
+    semi: ['error', 'always'],
+    'dot-location': ['error', 'property'],
+    eqeqeq: ['error', 'always'],
+    // React
+    'react/jsx-filename-extension': 0,
+    'react/jsx-indent': [2, 2, { checkAttributes: true, indentLogicalExpressions: true }],
+    'react/no-unescaped-entities': 0,
+    // Prettier
+    'prettier/prettier': ['error', prettierOptions],
+    // Import
+    'import/no-extraneous-dependencies': 0,
+    'import/no-unresolved': [0, { caseSensitive: false }],
+    // TS
+    '@typescript-eslint/explicit-member-accessibility': 0,
+    '@typescript-eslint/explicit-function-return-type': 0,
+    '@typescript-eslint/triple-slash-reference': 0,
+    '@typescript-eslint/no-var-requires': 0,
+  },
 
-	parser: '@typescript-eslint/parser',
+  settings: {
+    react: {
+      pragma: 'React',
+      version: 'detect',
+    },
+  },
 };
