+++ templates/typescript/.gitattributes
@@ -101,7 +101,7 @@ TODO              text
 *.conf            text
 *.config          text
 .editorconfig     text
-.env              text
+# .env              text
 .gitattributes    text
 .gitconfig        text
 .htaccess         text
