+++ templates/typescript/.prettierignore
@@ -1,6 +1,6 @@
 build/
+dist/
 node_modules/
 package-lock.json
 yarn.lock
-package.json
 coverage
