+++ templates/typescript/babel.config.js
@@ -1,19 +1,19 @@
 module.exports = {
-	presets: [
-		['@babel/preset-env', { modules: false }],
-		'@babel/preset-typescript',
-		'@babel/preset-react',
-	],
+  presets: [
+    ['@babel/preset-env', { modules: false }],
+    '@babel/preset-typescript',
+    '@babel/preset-react',
+  ],
 
-	plugins: [
-		'react-hot-loader/babel',
-		'@babel/plugin-proposal-class-properties',
-		'@babel/plugin-syntax-dynamic-import',
-	],
+  plugins: [
+    'react-hot-loader/babel',
+    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-syntax-dynamic-import',
+  ],
 
-	env: {
-		test: {
-			plugins: ['@babel/plugin-transform-modules-commonjs', 'dynamic-import-node'],
-		},
-	},
+  env: {
+    test: {
+      plugins: ['@babel/plugin-transform-modules-commonjs', 'dynamic-import-node'],
+    },
+  },
 };
