+++ templates/typescript/husky.config.js
@@ -1,0 +1,7 @@
+const tasks = (arr) => arr.join(' && ');
+
+module.exports = {
+  hooks: {
+    'pre-commit': tasks(['yarn lint', 'yarn precise-commits']),
+  },
+};
