+++ templates/typescript/jest.config.js
@@ -1,58 +1,58 @@
 module.exports = {
-	roots: ['<rootDir>/src'],
+  roots: ['<rootDir>/src'],
 
-	verbose: true,
+  verbose: true,
 
-	transform: {
-		'^.+\\.js$': 'babel-jest',
-		'.(ts|tsx)': 'ts-jest',
-		'^.+\\.(css|scss|less)$': 'jest-css-modules',
-	},
+  transform: {
+    '^.+\\.js$': 'babel-jest',
+    '.(ts|tsx)': 'ts-jest',
+    '^.+\\.(css|scss|less)$': 'jest-css-modules',
+  },
 
-	testURL: 'http://localhost',
+  testURL: 'http://localhost',
 
-	testMatch: [
-		'<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}',
-		'<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}',
-	],
+  testMatch: [
+    '<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}',
+    '<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}',
+  ],
 
-	testPathIgnorePatterns: ['/node_modules/'],
+  testPathIgnorePatterns: ['/node_modules/'],
 
-	setupFiles: ['raf/polyfill', './jest/setupJest.js'],
+  setupFiles: ['raf/polyfill', './jest/setupJest.js'],
 
-	modulePaths: ['<rootDir>/src', '<rootDir>/node_modules'],
+  modulePaths: ['<rootDir>/src', '<rootDir>/node_modules'],
 
-	transformIgnorePatterns: [
-		'[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$',
-		'^.+\\.module\\.(css|sass|scss)$',
-	],
+  transformIgnorePatterns: [
+    '[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$',
+    '^.+\\.module\\.(css|sass|scss)$',
+  ],
 
-	moduleNameMapper: {
-		'^.+\\.(jpg|jpeg|png|gif|eot|webp|svg|ttf|woff|woff2)$': '<rootDir>/jest/fileMock.js',
-		'^.+\\.(css|sass|scss)$': 'identity-obj-proxy',
-		'^jest/(.*)$': '<rootDir>/jest/$1',
-	},
+  moduleNameMapper: {
+    '^.+\\.(jpg|jpeg|png|gif|eot|webp|svg|ttf|woff|woff2)$': '<rootDir>/jest/fileMock.js',
+    '^.+\\.(css|sass|scss)$': 'identity-obj-proxy',
+    '^jest/(.*)$': '<rootDir>/jest/$1',
+  },
 
-	moduleFileExtensions: [
-		'web.js',
-		'js',
-		'web.ts',
-		'ts',
-		'web.tsx',
-		'tsx',
-		'json',
-		'web.jsx',
-		'jsx',
-		'node',
-	],
+  moduleFileExtensions: [
+    'web.js',
+    'js',
+    'web.ts',
+    'ts',
+    'web.tsx',
+    'tsx',
+    'json',
+    'web.jsx',
+    'jsx',
+    'node',
+  ],
 
-	coverageReporters: ['html', 'text-summary', 'lcov'],
+  coverageReporters: ['html', 'text-summary', 'lcov'],
 
-	collectCoverage: true,
+  collectCoverage: true,
 
-	collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}', '!src/**/*.d.ts'],
+  collectCoverageFrom: ['src/**/*.{js,jsx,ts,tsx}', '!src/**/*.d.ts'],
 
-	watchPlugins: ['jest-watch-typeahead/filename', 'jest-watch-typeahead/testname'],
+  watchPlugins: ['jest-watch-typeahead/filename', 'jest-watch-typeahead/testname'],
 
-	watchPathIgnorePatterns: ['node_modules'],
+  watchPathIgnorePatterns: ['node_modules'],
 };
