+++ templates/typescript/package.json
@@ -14,16 +14,17 @@
     "analyze": false
   },
   "scripts": {
-    "start": "cross-env NODE_ENV=development webpack-dev-server --config ./webpack/webpack.development.js",
     "build": "cross-env NODE_ENV=production webpack --config ./webpack/webpack.production.js",
-    "lint": "eslint --ext .ts,.tsx,.js,.jsx --no-error-on-unmatched-pattern src/",
-    "lint:fix": "eslint --ext .ts,.tsx,.js,.jsx --no-error-on-unmatched-pattern --quiet --fix",
+    "check-types": "tsc",
+    "lint:fix": "eslint --fix",
+    "lint": "eslint src/",
+    "precise-commits": "precise-commits",
     "pretest": "yarn run test:clean && yarn run lint",
-    "test:clean": "rimraf ./coverage",
-    "test": "cross-env NODE_ENV=test jest --config jest.config.js --coverage",
-    "test:watch": "cross-env NODE_ENV=test jest --watchAll",
     "prettify": "prettier . --write",
-    "check-types": "tsc"
+    "start": "cross-env NODE_ENV=development webpack-dev-server --config ./webpack/webpack.development.js",
+    "test:clean": "rimraf ./coverage",
+    "test:watch": "cross-env NODE_ENV=test jest --watchAll",
+    "test": "cross-env NODE_ENV=test jest --config jest.config.js --coverage"
   },
   "dependencies": {
     "clean-webpack-plugin": "3.0.0",
@@ -47,11 +48,11 @@
     "@babel/preset-typescript": "7.10.4",
     "@types/enzyme": "3.10.5",
     "@types/jest": "26.0.10",
-    "@types/node": "14.6.0",
-    "@types/react": "16.9.46",
+    "@types/node": "14.6.1",
+    "@types/react": "16.9.48",
     "@types/react-dom": "16.9.8",
-    "@typescript-eslint/eslint-plugin": "3.10.0",
-    "@typescript-eslint/parser": "3.10.0",
+    "@typescript-eslint/eslint-plugin": "3.10.1",
+    "@typescript-eslint/parser": "3.10.1",
     "babel-jest": "26.3.0",
     "babel-loader": "8.1.0",
     "babel-plugin-dynamic-import-node": "2.3.3",
@@ -76,49 +77,35 @@
     "eslint-plugin-package-json": "0.1.4",
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-react": "7.20.6",
-		"eslint-plugin-react-hooks": "4.1.0",
-		"file-loader": "6.0.0",
+    "eslint-plugin-react-hooks": "4.1.0",
+    "file-loader": "6.0.0",
     "html-webpack-plugin": "^4.3.0",
+    "husky": "4.2.5",
     "identity-obj-proxy": "3.0.0",
     "jest": "26.4.2",
     "jest-css-modules": "2.1.0",
     "jest-enzyme": "7.1.2",
     "jest-watch-typeahead": "0.6.0",
-    "lint-staged": "10.2.11",
-    "mini-css-extract-plugin": "^0.10.0",
+    "mini-css-extract-plugin": "^0.11.0",
     "node-sass": "4.14.1",
     "optimize-css-assets-webpack-plugin": "5.0.3",
-    "prettier": "2.1.0",
+    "precise-commits": "1.0.2",
+    "prettier": "2.1.1",
     "raf": "3.4.1",
     "react-hot-loader": "4.12.21",
     "react-scripts": "3.4.3",
     "react-test-renderer": "16.13.1",
-    "sass-loader": "^10.0.0",
+    "sass-loader": "^10.0.1",
     "style-loader": "1.2.1",
     "stylelint": "13.6.1",
     "stylelint-config-sass-guidelines": "7.0.0",
     "terser-webpack-plugin": "4.1.0",
-    "ts-jest": "26.2.0",
-		"typescript-plugin-css-modules": "2.4.0",
-		"url-loader": "4.1.0",
+    "ts-jest": "26.3.0",
+    "typescript-plugin-css-modules": "2.4.0",
+    "url-loader": "4.1.0",
     "webpack-bundle-analyzer": "3.8.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
-  },
-  "lint-staged": {
-    "*.{ts,tsx,js,jsx}": [
-      "yarn run lint:fix",
-      "git add --force"
-    ],
-    "*.{md,json}": [
-      "prettier --write",
-      "git add --force"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npx lint-staged"
-    }
   },
   "repository": {
     "type": "git",
