+++ templates/typescript/public/manifest.json
@@ -13,13 +13,12 @@
       "sizes": "192x192"
     },
     {
-      "src": "logo512.png",
+      "src": "logo270.png",
       "type": "image/png",
-      "sizes": "512x512"
+      "sizes": "270x270"
     }
   ],
   "start_url": ".",
-  "display": "standalone",
   "theme_color": "#000000",
   "background_color": "#ffffff"
 }
