+++ templates/typescript/src/pages/App.tsx
