+++ templates/typescript/webpack/webpack.production.js
@@ -8,62 +8,62 @@ const { includePathFromSrc } = require('./paths');
 const { PUBLIC_PATH_PROD } = process.env;
 
 const webpackProduction = merge(webpackBaseConfig, {
-	entry: {
-		app: includePathFromSrc('index.ts'),
-	},
+  entry: {
+    app: includePathFromSrc('index.ts'),
+  },
 
-	output: {
-		filename: 'static/[name].[contenthash].js',
-		chunkFilename: 'static/[name].[contenthash].chunk.js',
-		publicPath: PUBLIC_PATH_PROD,
-	},
+  output: {
+    filename: 'static/[name].[contenthash].js',
+    chunkFilename: 'static/[name].[contenthash].chunk.js',
+    publicPath: PUBLIC_PATH_PROD,
+  },
 
-	mode: 'production',
+  mode: 'production',
 
-	devtool: 'source-map',
+  devtool: 'source-map',
 
-	performance: {
-		hints: 'error',
-		maxAssetSize: 150 * 1024, // 150 KiB
-		maxEntrypointSize: 150 * 1024, // 150 KiB
-	},
+  performance: {
+    hints: 'error',
+    maxAssetSize: 150 * 1024, // 150 KiB
+    maxEntrypointSize: 150 * 1024, // 150 KiB
+  },
 
-	plugins: [
-		new CompressionPlugin({
-			algorithm: 'gzip',
-			test: /\.js$|\.css$|\.html$/,
-			threshold: 10240,
-			minRatio: 0.8,
-		}),
-	],
+  plugins: [
+    new CompressionPlugin({
+      algorithm: 'gzip',
+      test: /\.js$|\.css$|\.html$/,
+      threshold: 10240,
+      minRatio: 0.8,
+    }),
+  ],
 
-	optimization: {
-		minimize: true,
+  optimization: {
+    minimize: true,
 
-		sideEffects: true,
+    sideEffects: true,
 
-		concatenateModules: true,
+    concatenateModules: true,
 
-		minimizer: [
-			new TerserPlugin({
-				terserOptions: {
-					warnings: false,
-					compress: {
-						comparisons: false,
-					},
-					parse: {},
-					mangle: true,
-					output: {
-						comments: false,
-						ascii_only: true,
-					},
-				},
-				sourceMap: true,
-			}),
+    minimizer: [
+      new TerserPlugin({
+        terserOptions: {
+          warnings: false,
+          compress: {
+            comparisons: false,
+          },
+          parse: {},
+          mangle: true,
+          output: {
+            comments: false,
+            ascii_only: true,
+          },
+        },
+        sourceMap: true,
+      }),
 
-			new OptimizeCSSAssetsPlugin(),
-		],
-	},
+      new OptimizeCSSAssetsPlugin(),
+    ],
+  },
 });
 
 module.exports = webpackProduction;
